Refactor MutationObserver
