+++ packages/child/index.js
@@ -57,11 +57,12 @@ import {
   warn,
 } from './console'
 import { getBoolean, getNumber } from './from-string'
+import createMutationObserver from './mutation'
 import createOverflowObserver from './overflow'
 import { PREF_END, PREF_START } from './perf'
 import { readFunction, readNumber, readString } from './read'
 import createResizeObserver from './resize'
-import visibilityObserver from './visibility'
+import createVisibilityObserver from './visibility'
 
 function iframeResizerChild() {
   const customCalcMethods = {
@@ -933,133 +934,32 @@ This version of <i>iframe-resizer</> can auto detect the most suitable ${type} c
 
   function setupVisibilityObserver() {
     log('Setup VisibilityObserver')
-    visibilityObserver(visibilityChange)
+    createVisibilityObserver(visibilityChange)
+  }
+
+  function updatePage(addedMutations, removedMutations) {
+    consoleEvent('updatePage')
+    applySelectors()
+    checkAndSetupTags()
+    checkOverflow()
+
+    for (const mutation of addedMutations) {
+      const elements = getAllElements(mutation)()
+      overflowObserver.attachObservers(elements)
+      resizeObserver.attachObserverToNonStaticElements(elements)
+    }
+
+    for (const mutation of removedMutations) {
+      const elements = getAllElements(mutation)()
+      overflowObserver.detachObservers(elements)
+      resizeObserver.detachObservers(elements)
+    }
+
+    endAutoGroup()
   }
 
   function setupMutationObserver() {
-    const addedMutations = new Set()
-    const removedMutations = new Set()
-    const newMutations = []
-    let pending = false
-    let perfMon = 0
-
-    const updateMutation = (mutations) => {
-      info('Mutations observed:', mutations)
-
-      for (const mutation of mutations) {
-        const { addedNodes, removedNodes } = mutation
-
-        for (const node of addedNodes) {
-          addedMutations.add(node)
-        }
-
-        for (const node of removedNodes) {
-          if (addedMutations.has(node)) {
-            addedMutations.delete(node)
-          } else {
-            removedMutations.add(node)
-          }
-        }
-      }
-    }
-
-    const DELAY = 16 // Corresponds to 60fps
-    const DELAY_MARGIN = 2
-    const DELAY_MAX = 200
-
-    let delayCount = 1
-
-    function updatePage(addedMutations, removedMutations) {
-      consoleEvent('updatePage')
-      applySelectors()
-
-      for (const mutation of addedMutations) {
-        const elements = getAllElements(mutation)()
-        overflowObserver.attachObservers(elements)
-        resizeObserver.attachObserverToNonStaticElements(elements)
-      }
-
-      for (const mutation of removedMutations) {
-        const elements = getAllElements(mutation)()
-        overflowObserver.detachObservers(elements)
-        resizeObserver.detachObservers(elements)
-      }
-      endAutoGroup()
-    }
-
-    function processMutations() {
-      const now = performance.now()
-      const delay = now - perfMon
-      const delayLimit = DELAY * delayCount++ + DELAY_MARGIN
-
-      // Back off if the callStack is busy with other stuff
-      if (delay > delayLimit && delay < DELAY_MAX) {
-        info('Backed off due to heavy workload on callStack')
-        log(
-          `Delay: %c${round(delay)}ms %c> Delay limit: %c${delayLimit}ms`,
-          HIGHLIGHT,
-          FOREGROUND,
-          HIGHLIGHT,
-        )
-        setTimeout(processMutations, DELAY * delayCount)
-        perfMon = now
-        return
-      }
-
-      delayCount = 1
-
-      newMutations.forEach(updateMutation)
-      newMutations.length = 0
-
-      updatePage(addedMutations, removedMutations)
-
-      // Rebuild elements lists for size calculation
-      checkAndSetupTags()
-      checkOverflow()
-
-      pending = false
-      addedMutations.clear()
-      removedMutations.clear()
-
-      sendSize(MUTATION_OBSERVER, 'Mutation Observed')
-    }
-
-    function mutationObserved(mutations) {
-      newMutations.push(mutations)
-      if (pending) return
-
-      perfMon = performance.now()
-      pending = true
-      requestAnimationFrame(processMutations)
-    }
-
-    function createMutationObserver() {
-      const observer = new window.MutationObserver(mutationObserved)
-      const target = document.querySelector('body')
-      const config = {
-        attributes: true,
-        attributeFilter: [IGNORE_ATTR, SIZE_ATTR],
-        attributeOldValue: false,
-        characterData: false,
-        characterDataOldValue: false,
-        childList: true,
-        subtree: true,
-      }
-
-      log('Setup <body> MutationObserver')
-      observer.observe(target, config)
-
-      return observer
-    }
-
-    const observer = createMutationObserver()
-
-    return {
-      disconnect() {
-        log('Disconnect MutationObserver')
-        observer.disconnect()
-      },
-    }
+    createMutationObserver(updatePage, sendSize)
   }
 
   function getMaxElement(side) {
